Add methods for retrieving wheel velocities for drive velocity PID tuner (fixes #33)

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/drive/DriveConstants.java
@@ -57,7 +57,7 @@ public class DriveConstants {
     );
 
 
-    public static double encoderTicksToInches(int ticks) {
+    public static double encoderTicksToInches(double ticks) {
         return WHEEL_RADIUS * 2 * Math.PI * GEAR_RATIO * ticks / TICKS_PER_REV;
     }
